Update manifest.json
This Github file has an error at the end. So when I try to deploy with git on my plesk it doesn't even load the setup page.
Conflict problem on the call of the CSS file.
I made a very slight correction that now makes it work.
This is my first contribution on github (and in open source in general) so please be indulgent if it is not done in the rules. Thank you ;)

Signed-off-by: HansGruber51 <122012968+HansGruber51@users.noreply.github.com>